requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
gunicorn==23.0.0
weasyprint==65.1
pandas==2.3.1
openpyxl==3.1.5
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
pytz==2025.2
selenium==4.18.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
webdriver-manager==4.0.1
PyYAML==6.0.1
WeasyPrint==61.2

# ═══════════════════════════════════════════════════════════════
# Dépendances Module AI
# ═══════════════════════════════════════════════════════════════
prophet==1.1.5
openai>=1.12.0
groq>=0.3.0

# ═══════════════════════════════════════════════════════════════
# Dépendances Réseau et HTTP
# ═══════════════════════════════════════════════════════════════
requests>=2.31.0

# ═══════════════════════════════════════════════════════════════
# Dépendances Imprimante USB (optionnel, seulement si imprimante locale)
# ═══════════════════════════════════════════════════════════════
# pyusb>=1.2.1  # Décommenter si imprimante USB locale nécessaire
